Attempted getting PDF export header/footer working fully
Can kind of get it working on DOMPDF but it's fairly fiddly in terms if sizing.
WKHTMLtoPDF takes a completely different approach to headers/footers,
Got it working to a degree but also needs different sizing and page control.
Not sure where to go on this.

Done some general re-org in this commit also, Might keep that.